+++ website/build.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-hugo
-lunr-hugo -i "content/**/*.md" -o ../docs/index.json -l toml